Move the resolve Request "mutator" to the RequestBuilder class
